Factored out grid API into separate file
