+++ src/sim/chunk/mod.rs
@@ -9,6 +9,7 @@ pub type ChunkFragment = [BlockData; CHUNK_SIZE];
 pub type ChunkArray = [[ChunkFragment; CHUNK_SIZE]; CHUNK_SIZE];
 
 pub mod region;
+pub mod grid;
 pub mod grid_tree;
 
 mod pos;
